+++ project_test/tests/test_002_parser.py
@@ -7,6 +7,8 @@ import pytest
 
 from rstview import parser
 
+print "FOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOO"
+print parser
 
 def test_parser_basic_content(settings):
     """Parse a basic input with default writer (html5) and returned body"""
@@ -49,6 +51,8 @@ def test_parser_invalid_syntax(settings):
     ("basic/input.rst", "basic/output-html4.html", "html4"),
     ("advanced/input.rst", "advanced/output-html5.html", "html5"),
     ("advanced/input.rst", "advanced/output-html4.html", "html4"),
+    ("invalid/input.rst", "invalid/output-html5.html", "html5"),
+    ("invalid/input.rst", "invalid/output-html4.html", "html4"),
 ])
 def test_parser_file(settings, storageparameters, source_filename, output_filename, writer):
     """Compare a rendered source to attempted return from a data fixture file"""
@@ -64,10 +68,11 @@ def test_parser_file(settings, storageparameters, source_filename, output_filena
     with open(output_filepath, 'r') as fp:
         attempted = fp.read()
 
-    ## Temporary
-    #parser.build_output(source, output_filepath, setting_key="default",
-                 #body_only=True, initial_header_level=None,
-                 #silent=False)
+    # Temporary
+    if 'invalid/' in source_filename:
+        parser.build_output(source, output_filepath, setting_key="default",
+                    body_only=True, initial_header_level=None,
+                    silent=False)
 
     render = parser.SourceParser(source, setting_key="default", body_only=True,
                                  initial_header_level=None, silent=False)