+++ rstview/views.py
@@ -2,6 +2,7 @@
 """
 ReStructuredText views
 """
+from django.conf import settings
 from django.views.generic import TemplateView
 from django.utils.safestring import mark_safe
 
@@ -12,26 +13,61 @@ class RSTFileView(TemplateView):
     """
     View to display a ReST file
 
-    Open the file, parse it and put his source and render into the context
+    Open given source filepath, parse it and render it.
+
+    Context will be expanded from some document related variables:
+
+    doc_title
+        The given document title.
+    doc_source
+        Source from given filepath.
+    doc_html
+        Rendered source from parser.
+    doc_parser_opts_name
+        Key name of parser options from
+        ``settings.RSTVIEW_PARSER_FILTER_SETTINGS``.
     """
     template_name = "rstview/fileview.html"
     doc_title = None
-    doc_file_path = None
-    doc_parser_kwargs = {
-        'silent':True,
-        'body_only':True
-    }
+    doc_path = None
+    doc_parser_silent = False
+    doc_parser_bodyonly = True
+    doc_parser_opts_name = 'default'
+
+    def get_document_title(self, **kwargs):
+        """Return document title from RSTFileView.doc_title"""
+        return self.doc_title
+
+    def get_parser_opts(self, **kwargs):
+        """Return parser options"""
+        return {
+            'setting_key': self.doc_parser_opts_name,
+            'silent': self.doc_parser_silent,
+            'body_only': self.doc_parser_bodyonly,
+        }
+
+    def get_source(self, **kwargs):
+        """Return file content"""
+        with open(self.doc_path,'r') as fp:
+            source = fp.read()
+        return source
+
+    def render_source(self, source, **kwargs):
+        """Parse given source and return result as safe for django template"""
+        output = parser.SourceParser(source, **self.get_parser_opts())
+
+        return mark_safe(output)
 
     def get_context_data(self, **kwargs):
         context = super(RSTFileView, self).get_context_data(**kwargs)
 
-        f = open(self.doc_file_path,'r')
-        source = f.read()
-        f.close()
+        source = self.get_source(**kwargs)
+        rendered = self.render_source(source, **kwargs)
 
         context.update({
-            'doc_title': self.doc_title,
+            'doc_title': self.get_document_title(**kwargs),
             'doc_source': source,
-            'doc_html': mark_safe(parser.SourceParser(source, **self.doc_parser_kwargs)),
+            'doc_html': rendered,
+            'doc_parser_opts_name': self.doc_parser_opts_name,
         })
         return context